# Feature/better defaults improves performance through better torch dataloader usage