--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -2,9 +2,9 @@ Package: tabnet
 Title: Fit 'TabNet' Models for Classification and Regression
 Version: 0.3.0.9000
 Authors@R: c(
-    person(given = "Daniel", family = "Falbel", role = c("aut", "cre"), email = "daniel@rstudio.com"),
+    person(given = "Daniel", family = "Falbel", role = c("aut"), email = "daniel@rstudio.com"),
     person(family = "RStudio", role = c("cph")),
-    person(given = "Christophe", family = "Regouby", role = c("ctb"))
+    person(given = "Christophe", family = "Regouby", role = c("cre"))
     )
 Description: Implements the 'TabNet' model by Sercan O. Arik et al (2019) <arXiv:1908.07442>
     and provides a consistent interface for fitting and creating predictions. It's