--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -4,7 +4,7 @@ Version: 0.3.0.9000
 Authors@R: c(
     person(given = "Daniel", family = "Falbel", role = c("aut"), email = "daniel@rstudio.com"),
     person(family = "RStudio", role = c("cph")),
-    person(given = "Christophe", family = "Regouby", role = c("cre"))
+    person(given = "Christophe", family = "Regouby", role = c("cre"), email = "christophe.regouby@free.fr")
     )
 Description: Implements the 'TabNet' model by Sercan O. Arik et al (2019) <arXiv:1908.07442>
     and provides a consistent interface for fitting and creating predictions. It's