--- a/R/explain.R
+++ b/R/explain.R
@@ -20,9 +20,9 @@
 #'
 #' n <- 1000
 #' x <- data.frame(
-#'   x = runif(n),
-#'   y = runif(n),
-#'   z = runif(n)
+#'   x = rnorm(n),
+#'   y = rnorm(n),
+#'   z = rnorm(n)
 #' )
 #'
 #' y <- x$x