--- a/R/model.R
+++ b/R/model.R
@@ -9,37 +9,34 @@
 #'
 #' @param x a data frame
 #' @param y a response vector
-resolve_data <- function(x, y) {
-  stopifnot("Error: found missing values in the response vector" = sum(is.na(y))==0)
+#' @param device torch_device to send the tensor to
+resolve_data <- function(x, y, device) {
+  cat_idx <- which(sapply(x, is.factor))
+  cat_dims <- sapply(cat_idx, function(i) nlevels(x[[i]]))
   # convert factors to integers
-  x_ <- x
-  for (v in seq_along(x_)) {
-    if (is.factor(x_[[v]]))
-      x_[[v]] <- as.numeric(x_[[v]])
-  }
-  x_tensor <- torch::torch_tensor(as.matrix(x_), dtype = torch::torch_float())
-  x_na_mask <- x %>% is.na %>% as.matrix %>% torch::torch_tensor(dtype = torch::torch_bool())
-  if (ncol(y) == 1)
-    y <- y[[1]]
+  x[,cat_idx] <- sapply(cat_idx, function(i) as.integer(x[[i]]))
+
+  x_tensor <- torch::torch_tensor(as.matrix(x), dtype = torch::torch_float(), device = device)
+  x_na_mask <- x %>% is.na %>% as.matrix %>% torch::torch_tensor(dtype = torch::torch_bool(), device = device)
 
   if (is.factor(y)) {
-    y_tensor <- torch::torch_tensor(as.integer(y), dtype = torch::torch_int64())
+    y_tensor <- torch::torch_tensor(as.integer(y), dtype = torch::torch_long(), device = device)
   } else {
-    y_tensor <- torch::torch_tensor(y, dtype = torch::torch_float())$unsqueeze(2)
+    y_tensor <- torch::torch_tensor(y, dtype = torch::torch_float(), device = device)$unsqueeze(2)
   }
 
-  cat_idx <- which(sapply(x, is.factor))
-  cat_dims <- sapply(cat_idx, function(i) length(levels(x[[i]])))
-
+  # TODO put that initialization  apart
   if (is.factor(y))
-    output_dim <- max(as.integer(y))
+    output_dim <- nlevels(y)
   else
-    output_dim <- 1
+    output_dim <- 1L
 
-  input_dim <- ncol(x)
+  input_dim <- torch::torch_tensor(ncol(x), device=device)
 
-  list(x = x_tensor, x_na_mask = x_na_mask, y = y_tensor, cat_idx = cat_idx, output_dim = output_dim,
-       input_dim = input_dim, cat_dims = cat_dims)
+  list(x = x_tensor, x_na_mask = x_na_mask, y = y_tensor,
+       cat_idx = torch::torch_tensor(cat_idx, device=device),
+       output_dim = torch::torch_tensor(output_dim, device=device),
+       input_dim = input_dim, cat_dims = torch::torch_tensor(cat_dims, device=device))
 }
 
 #' Configuration for TabNet models
@@ -214,17 +211,17 @@ resolve_early_stop_monitor <- function(early_stopping_monitor, valid_split) {
   early_stopping_monitor
 }
 
-
-batch_to_device <- function(batch, device) {
-  batch <- list(x = batch$x, na_mask=batch$na_mask, y  = batch$y)
-  lapply(batch, function(x) {
-    x$to(device = device)
-  })
-}
+#
+# batch_to_device <- function(batch, device) {
+#   batch <- list(x = batch$x, na_mask=batch$na_mask, y  = batch$y)
+#   lapply(batch, function(x) {
+#     x$to(device = device)
+#   })
+# }
 
 train_batch <- function(network, optimizer, batch, config) {
   # forward pass
-  output <- network(batch$x, batch$na_mask)
+  output <- network(batch$x, batch$x_na_mask)
   loss <- config$loss_fn(output[[1]], batch$y)
 
   # Add the overall sparsity loss
@@ -245,7 +242,7 @@ train_batch <- function(network, optimizer, batch, config) {
 
 valid_batch <- function(network, batch, config) {
   # forward pass
-  output <- network(batch$x, batch$na_mask)
+  output <- network(batch$x, batch$x_na_mask)
   loss <- config$loss_fn(output[[1]], batch$y)
 
   # Add the overall sparsity loss
@@ -277,6 +274,7 @@ tabnet_initialize <- function(x, y, config = tabnet_config()) {
   torch::torch_manual_seed(sample.int(1e6, 1))
   has_valid <- config$valid_split > 0
 
+
   if (config$device == "auto") {
     if (torch::cuda_is_available())
       device <- "cuda"
@@ -286,35 +284,53 @@ tabnet_initialize <- function(x, y, config = tabnet_config()) {
     device <- config$device
   }
 
+  # simplify y into vector
+  if (!is.atomic(y)) {
+    # currently not supporting multilabel
+    y <- y[[1]]
+  }
+
   if (has_valid) {
     n <- nrow(x)
     valid_idx <- sample.int(n, n*config$valid_split)
+    valid_x <- x[valid_idx, ]
 
-    if (is.data.frame(y)) {
-      valid_y <- y[valid_idx,]
-      train_y <- y[-valid_idx,]
-    } else if (is.numeric(y) || is.factor(y)) {
+    # if (is.data.frame(y)) {
+    #   # currently not supporting multilabel
+    #   valid_y <- y[valid_idx,][[1]]
+    #   train_y <- y[-valid_idx,][[1]]
+    # } else if (is.numeric(y) || is.factor(y)) {
       valid_y <- y[valid_idx]
       train_y <- y[-valid_idx]
-    }
+    # }
 
-    valid_lst <- list(x = x[valid_idx, ], na_mask = x[valid_idx, ] %>% is.na, y = valid_y)
+    valid_ds <-   torch::dataset(
+      initialize = function() {},
+      .getbatch = function(batch) {resolve_data(valid_x[batch,], valid_y[batch], device=device)},
+      .length = function() {nrow(valid_x)}
+    )()
     x <- x[-valid_idx, ]
     y <- train_y
   }
 
-  # training matrix
-  train_mat <- resolve_data(x, y)
+  # training dataset
+  train_ds <-   torch::dataset(
+    initialize = function() {},
+    .getbatch = function(batch) {resolve_data(x[batch,], y[batch], device=device)},
+    .length = function() {nrow(x)}
+  )()
+  # we can get training_set parameters from the 2 first samples
+  train <- train_ds$.getbatch(batch = c(1:2))
 
   # resolve loss
-  config$loss_fn <- resolve_loss(config$loss, train_mat$y$dtype)
+  config$loss_fn <- resolve_loss(config$loss, train$y$dtype)
 
   # create network
   network <- tabnet_nn(
-    input_dim = train_mat$input_dim,
-    output_dim = train_mat$output_dim,
-    cat_idxs = train_mat$cat_idx,
-    cat_dims = train_mat$cat_dims,
+    input_dim = as.integer(train$input_dim$to(device="cpu")),
+    output_dim = as.integer(train$output_dim$to(device="cpu")),
+    cat_idxs = as.integer(train$cat_idx$to(device="cpu")),
+    cat_dims = as.integer(train$cat_dims$to(device="cpu")),
     n_d = config$n_d,
     n_a = config$n_a,
     n_steps = config$n_steps,
@@ -359,29 +375,46 @@ tabnet_train_supervised <- function(obj, x, y, config = tabnet_config(), epoch_s
     device <- config$device
   }
 
+  # simplify y into vector
+  if (!is.atomic(y)) {
+    # currently not supporting multilabel
+    y <- y[[1]]
+  }
+
   # dataset to dataloaders
   has_valid <- config$valid_split > 0
   if (has_valid) {
     n <- nrow(x)
     valid_idx <- sample.int(n, n*config$valid_split)
+    valid_x <- x[valid_idx, ]
 
-    if (is.data.frame(y)) {
-      valid_y <- y[valid_idx,]
-      train_y <- y[-valid_idx,]
-    } else if (is.numeric(y) || is.factor(y)) {
+    # if (is.data.frame(y)) {
+    #   # currently not supporting multilabel
+    #   valid_y <- y[valid_idx,][[1]]
+    #   train_y <- y[-valid_idx,][[1]]
+    # } else if (is.numeric(y) || is.factor(y)) {
       valid_y <- y[valid_idx]
       train_y <- y[-valid_idx]
-    }
-
-    valid_lst <- list(x = x[valid_idx, ], na_mask = x[valid_idx, ] %>% is.na, y = valid_y)
+    # }
+    # validation dataset
+    valid_ds <-   torch::dataset(
+      initialize = function() {},
+      .getbatch = function(batch) {resolve_data(valid_x[batch,], valid_y[batch], device=device)},
+      .length = function() {nrow(valid_x)}
+    )()
     x <- x[-valid_idx, ]
     y <- train_y
   }
 
-  # training data
-  train_mat <- resolve_data(x, y)
+  # training dataset
+  train_ds <-   torch::dataset(
+    initialize = function() {},
+    .getbatch = function(batch) {resolve_data(x[batch,], y[batch], device=device)},
+    .length = function() {nrow(x)}
+  )()
+
   train_dl <- torch::dataloader(
-    torch::tensor_dataset(x = train_mat$x, na_mask = train_mat$x_na_mask, y = train_mat$y),
+    train_ds,
     batch_size = config$batch_size,
     drop_last = config$drop_last,
     shuffle = TRUE ,
@@ -390,9 +423,9 @@ tabnet_train_supervised <- function(obj, x, y, config = tabnet_config(), epoch_s
 
   # validation data
   if (has_valid) {
-    valid_mat <- resolve_data(valid_lst$x, valid_lst$y)
+    # valid_mat <- resolve_data(valid_lst$x, valid_lst$y, device=device)
     valid_dl <- torch::dataloader(
-      torch::tensor_dataset(x = valid_mat$x, na_mask = valid_mat$x_na_mask, y = valid_mat$y),
+      valid_ds,
       batch_size = config$batch_size,
       drop_last = FALSE,
       shuffle = FALSE ,
@@ -401,7 +434,7 @@ tabnet_train_supervised <- function(obj, x, y, config = tabnet_config(), epoch_s
   }
 
   # resolve loss
-  config$loss_fn <- resolve_loss(config$loss, train_mat$y$dtype)
+  config$loss_fn <- resolve_loss(config$loss, train_ds$.getbatch(batch = c(1:2))$y$dtype)
 
   # restore network from model and send it to device
   network <- obj$fit$network
@@ -453,7 +486,7 @@ tabnet_train_supervised <- function(obj, x, y, config = tabnet_config(), epoch_s
       )
 
     coro::loop(for (batch in train_dl) {
-      m <- train_batch(network, optimizer, batch_to_device(batch, device), config)
+      m <- train_batch(network, optimizer, batch, config)
       if (config$verbose) pb$tick(tokens = m)
       train_metrics <- c(train_metrics, m)
     })
@@ -468,7 +501,7 @@ tabnet_train_supervised <- function(obj, x, y, config = tabnet_config(), epoch_s
     network$eval()
     if (has_valid) {
       coro::loop(for (batch in valid_dl) {
-        m <- valid_batch(network, batch_to_device(batch, device), config)
+        m <- valid_batch(network, batch, config)
         valid_metrics <- c(valid_metrics, m)
       })
       metrics[[epoch]][["valid"]] <- transpose_metrics(valid_metrics)
@@ -520,19 +553,22 @@ tabnet_train_supervised <- function(obj, x, y, config = tabnet_config(), epoch_s
   network$to(device = "cpu")
 
   importance_sample_size <- config$importance_sample_size
-  if (is.null(config$importance_sample_size) && train_mat$x$shape[1] > 1e5) {
-    rlang::warn(c(glue::glue("Computing importances for a dataset with size {train_mat$x$shape[1]}."),
+  if (is.null(config$importance_sample_size) && train_ds$.length() > 1e5) {
+    rlang::warn(c(glue::glue("Computing importances for a dataset with size {train_ds$.length()}."),
                 "This can consume too much memory. We are going to use a sample of size 1e5",
                 "You can disable this message by using the `importance_sample_size` argument."))
     importance_sample_size <- 1e5
   }
-  indexes <- torch::torch_randint(
-    1, train_mat$x$shape[1], min(importance_sample_size, train_mat$x$shape[1]),
+  indexes <- as.numeric(torch::torch_randint(
+    1, train_ds$.length(), min(importance_sample_size, train_ds$.length()),
     dtype = torch::torch_long()
-  )
+  ))
   importances <- tibble::tibble(
     variables = colnames(x),
-    importance = compute_feature_importance(network, train_mat$x[indexes,..], train_mat$x_na_mask[indexes,..])
+    importance = compute_feature_importance(
+      network,
+      train_ds$.getbatch(batch =indexes)$x$to(device = "cpu"),
+      train_ds$.getbatch(batch =indexes)$x_na_mask$to(device = "cpu"))
   )
 
   list(
@@ -545,7 +581,13 @@ tabnet_train_supervised <- function(obj, x, y, config = tabnet_config(), epoch_s
 }
 
 predict_impl <- function(obj, x, batch_size = 1e5) {
-  predict_mat <- resolve_data(x, y = data.frame(rep(1, nrow(x))))
+  # prediction dataset
+  device = obj$fit$config$device
+  predict_ds <-   torch::dataset(
+    initialize = function() {},
+    .getbatch = function(batch) {resolve_data(x[batch,], rep(1, nrow(x)), device=device)},
+    .length = function() {nrow(x)}
+  )()
 
   network <- obj$fit$network
   num_workers <- obj$fit$config$num_workers
@@ -553,14 +595,15 @@ predict_impl <- function(obj, x, batch_size = 1e5) {
   network$eval()
 
   predict_dl <- torch::dataloader(
-    torch::tensor_dataset(x = predict_mat$x, na_mask = predict_mat$x_na_mask),
+    predict_ds,
     batch_size = batch_size,
     drop_last = FALSE,
     shuffle = FALSE ,
-    num_workers = num_workers
+    # num_workers = num_workers
+    num_workers = 0L
   )
   coro::loop(for (batch in predict_dl) {
-    yhat <- c(yhat, network(batch$x, batch$na_mask)[[1]])
+    yhat <- c(yhat, network(batch$x, batch$x_na_mask)[[1]])
   })
 
   torch::torch_cat(yhat)