--- a/R/model.R
+++ b/R/model.R
@@ -268,10 +268,11 @@ tabnet_initialize <- function(x, y, config = tabnet_config()) {
 
 
   if (config$device == "auto") {
-    if (torch::cuda_is_available())
+    if (torch::cuda_is_available()){
       device <- "cuda"
-    else
+    } else {
       device <- "cpu"
+    }
   } else {
     device <- config$device
   }
@@ -351,10 +352,11 @@ tabnet_train_supervised <- function(obj, x, y, config = tabnet_config(), epoch_s
   torch::torch_manual_seed(sample.int(1e6, 1))
 
   if (config$device == "auto") {
-    if (torch::cuda_is_available())
+    if (torch::cuda_is_available()){
       device <- "cuda"
-    else
+    } else {
       device <- "cpu"
+    }
   } else {
     device <- config$device
   }