… .github/workflows/R-CMD-check.yaml
… .github/workflows/test-coverage.yaml
… NEWS.md
… R/tab-network.R
… man/resolve_data.Rd
… man/tabnet.Rd
… man/tabnet_config.Rd
… tests/testthat.R
… tests/testthat/_snaps/hardhat_interfaces.md
… tests/testthat/_snaps/pretraining.md
… tests/testthat/setup.R
… tests/testthat/test-explain.R
… tests/testthat/test-hardhat.R
… tests/testthat/test-hardhat_interfaces.R
… tests/testthat/test-hardhat_parameters.R
… tests/testthat/test-hardhat_scenarios.R
… tests/testthat/test-missing_values.R
… tests/testthat/test-model.R
… tests/testthat/test-plot.R
… tests/testthat/test-pretraining.R
… tests/testthat/test-resolve_data.R
… tests/testthat/test_resolve_data.R
… vignettes/Missing_data_predictors.Rmd
… vignettes/ames_fit.png
… vignettes/ames_fit_vip_.png
… vignettes/ames_mas_vnr_hist.png
… vignettes/ames_missing_fit.png
… vignettes/ames_missing_fit_vip.png
… vignettes/ames_missing_pretrain.png
… vignettes/ames_missing_pretrain_vip.png