This is an improvement step in usability and performance with
- larger `batch_size` by default for a 70% performance gain
- move `resolve_data` into datalaoder and dataset preprocessing and thus
- Fix #78 
- Fix #71 
- Fix #82 (as per test result)
- use global test fixture to save time for tests
- improve readability of early-stopping code
- Recorded performance on my laptop for #79 (all are times in sec. , the lower, the better) 

|num_workers | CPU| GPU
|---|---|---|
|0L|134s| 46s|
|1L| | 105s|
|4L|227s| |

Conclusion : as expected, 4 workers is not enough to accelerate preprocessing. (note: `num_workers` with GPU is limited here because my GPU cannot afford more than two R process in memory)